Rename folder
Every time I forget python hates hyphenated things. This fixes that.
